+++ packages/io-wallet-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-wallet-common
 
+## 1.4.2
+
+### Patch Changes
+
+- c8e7082: Updated node version and migrated from yarn to pnpm
+
+## 1.4.1
+
+### Patch Changes
+
+- f7f674d: Strengthened data validation in createWalletInstance and updated WalletInstance schema to allow string osPatchLevel for revoked instances
+
 ## 1.4.0
 
 ### Minor Changes